--- .env
@@ -1,1 +1,0 @@
-REACT_APP_GOOGLE_CLIENT_ID=https://cse403-sp22.slack.com/files/U038WKWQMFX/F03DT0RQELQ/client_secret_215324184605-v64ebk4kddt6aufdkmbaf0iv38doeqdb.apps.googleusercontent.com.json?origin_team=T039K0BN264&origin_channel=C03BB9KDGPJ
