+++ main/front-end/.env
@@ -1,0 +1,1 @@
+REACT_APP_GOOGLE_CLIENT_ID=205317701531-od80vq0biekitm7bq8irtfoen3fhpfo0.apps.googleusercontent.com
