+++ main/front-end/public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 
 <head>
-	<meta charset="utf-8" />
+	<!-- <meta charset="utf-8" />
 	<meta name="viewport"
 		content="width=device-width, initial-scale=1" />
 	<meta name="theme-color"
@@ -15,7 +15,10 @@
 		integrity=
 "sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
 		crossorigin="anonymous">
-	<title>React App</title>
+	<title>React App</title> -->
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Pacifico&display=swap" rel="stylesheet">
 </head>
 
 <body>
